Shop_Danawa/source_connector/requirements.txt
requests
kafka-python
mysql-connector-python
